refactor(wework): 重构企业微信配置的日志记录方式
- 将日志记录从 WeWorkProperties 移至 WeWorkAutoConfiguration
- 添加了脱敏方法 maskCorpId 来保护企业ID隐私
- 在配置类中统一进行功能启用日志记录
- 移除了属性类中的 PostConstruct 初始化方法
- 统一了日志输出
