+++ junoyi-framework/junoyi-framework-wechat/src/main/java/com/junoyi/wework/config/WxMaConfiguration.java
@@ -5,6 +5,7 @@ import cn.binarywang.wx.miniapp.api.impl.WxMaServiceImpl;
 import cn.binarywang.wx.miniapp.config.impl.WxMaDefaultConfigImpl;
 import com.junoyi.wework.properties.WxMpProperties;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
@@ -18,6 +19,7 @@ import org.springframework.context.annotation.Configuration;
  *
  * @author Fan
  */
+@Slf4j
 @Configuration
 @RequiredArgsConstructor
 @EnableConfigurationProperties(WxMpProperties.class)
@@ -33,6 +35,8 @@ public class WxMaConfiguration {
      */
     @Bean
     public WxMaService wxMaService() {
+        log.info("微信小程序功能已启用 [AppID: {}]", maskAppId(wxMpProperties.getAppId()));
+
         WxMaDefaultConfigImpl config = new WxMaDefaultConfigImpl();
         config.setAppid(wxMpProperties.getAppId());
         config.setSecret(wxMpProperties.getSecret());
@@ -41,6 +45,16 @@ public class WxMaConfiguration {
         service.setWxMaConfig(config);
 
         return service;
+    }
+
+    /**
+     * 脱敏 AppID，只显示前4位和后4位
+     */
+    private String maskAppId(String appId) {
+        if (appId == null || appId.length() <= 8) {
+            return "****";
+        }
+        return appId.substring(0, 4) + "****" + appId.substring(appId.length() - 4);
     }
 }
