+++ junoyi-framework/junoyi-framework-wechat/src/main/java/com/junoyi/wework/properties/WxMpProperties.java
@@ -1,8 +1,6 @@
 package com.junoyi.wework.properties;
 
-import jakarta.annotation.PostConstruct;
 import lombok.Data;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
@@ -14,7 +12,6 @@ import org.springframework.stereotype.Component;
  *
  * @author Fan
  */
-@Slf4j
 @Data
 @Component
 @ConfigurationProperties(prefix = "junoyi.wechat.mp")
@@ -34,23 +31,4 @@ public class WxMpProperties {
      * 小程序 AppSecret
      */
     private String secret;
-
-    @PostConstruct
-    public void init() {
-        if (Boolean.TRUE.equals(enabled)) {
-            log.info("WeChat mini program function is enabled [AppID: {}]", maskAppId(appId));
-        } else {
-            log.info("WeChat mini program function is not enabled");
-        }
-    }
-
-    /**
-     * 脱敏 AppID，只显示前4位和后4位
-     */
-    private String maskAppId(String appId) {
-        if (appId == null || appId.length() <= 8) {
-            return "****";
-        }
-        return appId.substring(0, 4) + "****" + appId.substring(appId.length() - 4);
-    }
 }
