+++ junoyi-framework/junoyi-framework-wechat/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,3 @@
+com.junoyi.wework.config.WxMaConfiguration
+com.junoyi.wework.config.WxPayConfiguration
+
